requirements.txt
click==6.6
Django==1.10.2
django-cors-headers==1.2.2
djangorestframework==3.4.7
Flask==0.11.1
Flask-Cors==3.0.2
hipchat==0.0.1
hiplogging==2.0.0
hypchat==0.21
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
phonenumbers==7.5.2
python-dateutil==2.5.3
requests==2.11.1
setupfiles==0.0.13
six==1.10.0
Werkzeug==0.11.11
